+++ plans.h
@@ -46,8 +46,9 @@ struct FlawSelectionOrder;
  * Causal link.
  */
 struct Link {
+
     /* Constructs a causal link. */
-    Link(size_t from_id, StepTime effect_time, const OpenCondition& open_cond);
+    Link(int from_id, StepTime effect_time, const OpenCondition& open_cond);
 
     /* Constructs a causal link. */
     Link(const Link& l);
@@ -56,13 +57,13 @@ struct Link {
     ~Link();
 
     /* Returns the id of step that link goes from. */
-    size_t from_id() const { return from_id_; }
+    int from_id() const { return from_id_; }
 
     /* Returns the time of effect satisfying link. */
     StepTime effect_time() const { return effect_time_; }
 
     /* Returns the id of step that link goes to. */
-    size_t to_id() const { return  to_id_; }
+    int to_id() const { return to_id_; }
 
     /* Returns the condition satisfied by link. */
     const Literal& condition() const { return *condition_; }
@@ -71,15 +72,20 @@ struct Link {
     FormulaTime condition_time() const { return condition_time_; }
 
 private:
-    /* Id of step that link goes from. */
-    size_t from_id_;
-    /* Time of effect satisfying link. */
+    
+	/* Id of step that link goes from. */
+    int from_id_;
+    
+	/* Time of effect satisfying link. */
     StepTime effect_time_;
-    /* Id of step that link goes to. */
-    size_t to_id_;
-    /* Condition satisfied by link. */
+    
+	/* Id of step that link goes to. */
+    int to_id_;
+    
+	/* Condition satisfied by link. */
     const Literal* condition_;
-    /* Time of condition satisfied by link. */
+    
+	/* Time of condition satisfied by link. */
     FormulaTime condition_time_;
 };
